+++ changelog.md
@@ -71,9 +71,12 @@ with visibility of the Dataset.
   - Link to Dataset details page for each dataset
 * Remove unused ajax POST function.
 * Remove unused JS variables.
+* Users must now be the author of a Scan Report or an admin of the parent Dataset
+to change the author of a Scan Report.
 * Can now include a datasets query in the projects list endpoint to include datasets in the result
 * Improved django admin pages' responsiveness and informativeness
 * Users who are not admins/editors of a scan report table cannot edit the form data.
+* User is now automatically added as admin to Datasets they create.
 
 ## v1.4.0 was released 02/02/22